requirements.txt
torch
transformers
hydra-core
omegaconf
Pillow
accelerate
nvidia-ml-py
langchain-core
datasets
sacrebleu
unbabel-comet
bert-score
nltk
psutil
compressed_tensors
pandas
seaborn
huggingface_hub[hf_xet]